Move more shared options methods to separate objects
